+++ application/external-api/src/main/java/com/namo/spring/application/external/api/group/dto/MeetingScheduleRequest.java
@@ -83,11 +83,4 @@ public class MeetingScheduleRequest {
 		private Long meetingScheduleId;
 		private List<Integer> alarmDates;
 	}
-
-	@Getter
-	@AllArgsConstructor
-	@NoArgsConstructor
-	public static class PostMeetingScheduleTextDto {
-		private String text;
-	}
 }
